requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
jupyter
matplotlib
seaborn
scipy
requests>=2.31.0
openai>=1.0.0
plotly>=5.0.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
